~ refactored eq processor
